clean up the code [move context out of App]
